Anadida vista gestion_proveedores.html incompleta

+++ TelExpress/src/main/java/com/example/telexpress/repository/ProveedorRepository.java
@@ -1,10 +1,10 @@
 package com.example.telexpress.repository;
 
-import com.example.telexpress.entity.Producto;
+
 import com.example.telexpress.entity.Proveedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProveedorRepository extends JpaRepository<Proveedor,Integer> {
+public interface ProveedorRepository extends JpaRepository<Proveedor,Long> {
 }
